export helper functions to utils module
